requirements.txt
google-auth==2.12.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
google-api-python-client==2.46.0
requests==2.26.0
sqlite3
